backend/requirements.txt
fastapi[standard]==0.115.12
sqlalchemy[asyncio]==2.0.38
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
uvicorn[standard]==0.34.0

# Dev / test
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
ruff>=0.9.0
